Move KibanaMigrator into Server SavedObjectsService (#43433)
* Rename SavedObjectsService -> SavedObjectsLegacyService

* Expose legacy pluginSpecs from Core LegacyService

* Expose legacy uiExports from Core LegacyService

* Move kibana config to NP

* Expose pluginExtendedConfig from LegacyService

* Make KibanaMigrator NP compatible

* KibanaMigrator -> NP SavedObjectsService

* SavedObjectsService never stop retrying ES connection error

* Move waiting for migrations to complete till after legacy service start

* Fix ESArchiver's KibanaMigrator

* Fix reload logging config tests

* Run migrations on savedobjects start

* Fix env tests

* Fix and make legacy tests more robust/isolated

* Cleanup code

* Fix invalid config test

* Fix SavedObject Migrations logging test

* SavedObjectsService tests

* Lifecycle logging and improve getting kibanaConfig instance

* Fix awaitMigration bug and test

* Fix typing error

* Review comments

* Remove unecessary KibanaConfig class

* Move legacy plugin config extension, specs, uiExports entirely into Core

uiExports, specs, disabledSpecs, config now get injected into KbnServer

* Fix config deprecation test

* Use existing logger mock

* Create SavedObjectsConfig for migration config

* Define KibanaMigratorContract type

* KibanaMigratorContract -> IKibanaMigrator + docs improvements

* Fix esArchiver's KibanaMigrator

* Fix plugin generator integration test

* ConfigServiceContract -> IConfigService

* Address review comments

* Review nits

* Document migrations.skip config

* Review comments continued...

* awaitMigrations -> runMigrations

* Type improvements

+++ src/cli/serve/integration_tests/reload_logging_config.test.js
@@ -83,7 +83,7 @@ describe('Server logging configuration', function () {
     it('should be reloadable via SIGHUP process signaling', async function () {
       expect.assertions(3);
 
-      child = spawn(process.execPath, [kibanaPath, '--config', testConfigFile, '--oss'], {
+      child = spawn(process.execPath, [kibanaPath, '--config', testConfigFile, '--oss', '--verbose'], {
         stdio: 'pipe'
       });
 
@@ -114,7 +114,9 @@ describe('Server logging configuration', function () {
               const data = JSON.parse(line);
               sawJson = true;
 
-              if (data.tags.includes('listening')) {
+              // We know the sighup handler will be registered before
+              // root.setup() is called
+              if (data.message.includes('setting up root')) {
                 isJson = false;
                 setLoggingJson(false);
 
@@ -128,10 +130,9 @@ describe('Server logging configuration', function () {
               // the switch yet, so we ignore before switching over.
             } else {
               // Kibana has successfully stopped logging json, so kill the server.
-
               sawNonjson = true;
 
-              child.kill();
+              child && child.kill();
               child = undefined;
             }
           })
@@ -178,10 +179,11 @@ describe('Server logging configuration', function () {
         '--config', testConfigFile,
         '--logging.dest', logPath,
         '--plugins.initialize', 'false',
-        '--logging.json', 'false'
+        '--logging.json', 'false',
+        '--verbose'
       ]);
 
-      watchFileUntil(logPath, /http server running/, 2 * minute)
+      watchFileUntil(logPath, /starting server/, 2 * minute)
         .then(() => {
           // once the server is running, archive the log file and issue SIGHUP
           fs.renameSync(logPath, logPathArchived);
@@ -190,8 +192,8 @@ describe('Server logging configuration', function () {
         .then(() => watchFileUntil(logPath, /Reloaded logging configuration due to SIGHUP/, 10 * second))
         .then(contents => {
           const lines = contents.toString().split('\n');
-          // should be the first and only new line of the log file
-          expect(lines).toHaveLength(2);
+          // should be the first line of the new log file
+          expect(lines[0]).toMatch(/Reloaded logging configuration due to SIGHUP/);
           child.kill();
         })
         .then(done, done);
